dom/base/test/test_bug330925.html

<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=330925
-->
<head>
  <title>Test for Bug 330925</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=330925">Mozilla Bug 330925</a>

<p id="display">
<el-custom id="t"></el-custom>
</p>

<pre id="test">
<script class="testbody" type="text/javascript">
customElements.define("el-custom", class extends HTMLElement {
  constructor() {
    super();
    this.attachShadow({ mode: "open" });
    this.shadowRoot.innerHTML = `
xbl textnode1
<div>
xbl textnode2
<slot></slot>
</div>
    `;
  }
});

let t = document.getElementById("t");
let d = SpecialPowers.wrap(document);

let tn1 = t.shadowRoot.firstChild;
let tn2 = t.shadowRoot.firstChild.nextSibling.firstChild;

is(SpecialPowers.unwrap(d.getBindingParent(tn1)),
   t,
   "Wrong binding parent for shadow DOM node");

is(SpecialPowers.unwrap(d.getBindingParent(tn2)),
   t,
   "Wrong binding parent for child of shadow DOM node");

is(d.getBindingParent(t),
   null,
   "Non shadow DOM node should not have a binding parent");

is(d.getBindingParent(document.documentElement),
   null,
   "Document element should not have a binding parent");
</script>
</pre>
</body>
</html>
